requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.8
pycparser==2.19
pyOpenSSL==19.1.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
